refactor: 플랫폼 연동 해제·재연동 mutation 훅 분리
